ml_backend/requirements.txt
fastapi
uvicorn
pillow
pytesseract
google-generativeai
python-dotenv